server/requirements.txt
arxiv==2.1.3
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.2.0
dnspython==2.7.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
Flask-WTF==1.2.2
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pillow==11.1.0
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
sympy==1.13.3
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
PyJWT
